+++ sources/functional-administration/functional-administration-profile/src/test/java/fr/gouv/vitam/functional/administration/profile/api/impl/ProfileServiceImplTest.java
@@ -405,7 +405,7 @@ public class ProfileServiceImplTest {
         assertThat(one.getName()).isEqualTo(acm.getName());
 
         assertThat(one.getTenant()).isNotNull();
-        assertThat(one.getTenant()).isEqualTo(Long.valueOf(TENANT_ID));
+        assertThat(one.getTenant()).isEqualTo(Integer.valueOf(TENANT_ID));
 
     }
